requirements.txt
boto3==1.26.25
botocore==1.29.25
cachelib==0.9.0
click==8.1.3
dash==2.7.0
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-daq==0.5.0
dash-extensions==0.1.8
dash-html-components==2.0.0
dash-table==5.0.0
EditorConfig==0.12.3
Flask==2.2.2
Flask-Caching==2.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsbeautifier==1.14.7
Markdown==3.4.1
MarkupSafe==2.1.1
more-itertools==8.14.0
numpy==1.23.5
pandas==1.5.2
plotly==5.11.0
psutil==5.9.4
# pygame==2.1.2
python-dateutil==2.8.2
pytz==2022.6
s3transfer==0.6.0
six==1.16.0
tenacity==8.1.0
urllib3==1.26.13
Werkzeug==2.2.2
